refactor: extract backup shared types
