+++ mhm/src-tauri/src/backup.rs
@@ -2,35 +2,21 @@ use chrono::NaiveDateTime;
 use chrono::Utc;
 use serde::Serialize;
 use std::{
-    fmt, fs, io,
+    fs, io,
     path::{Path, PathBuf},
     sync::atomic::{AtomicBool, AtomicUsize, Ordering},
     time::Duration,
 };
 use tauri::{AppHandle, Emitter, Manager};
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum BackupReason {
-    Settings,
-    Checkout,
-    GroupCheckout,
-    NightAudit,
-    AppExit,
-    Manual,
-}
+mod types;
 
-impl BackupReason {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Settings => "settings",
-            Self::Checkout => "checkout",
-            Self::GroupCheckout => "group_checkout",
-            Self::NightAudit => "night_audit",
-            Self::AppExit => "app_exit",
-            Self::Manual => "manual",
-        }
-    }
-}
+pub use types::{
+    log_backup_request_error, BackupError, BackupOutcome, BackupPruneOutcome, BackupReason,
+    BackupRequestError,
+};
+#[allow(unused_imports)]
+pub use types::BackupRequestErrorKind;
 
 pub fn build_backup_filename(reason: BackupReason, timestamp: NaiveDateTime) -> String {
     format!(
@@ -42,102 +28,6 @@ pub fn build_backup_filename(reason: BackupReason, timestamp: NaiveDateTime) -> 
 
 pub fn is_managed_backup_file(name: &str) -> bool {
     parse_backup_filename(name).is_some()
-}
-
-#[allow(dead_code)]
-#[derive(Debug)]
-pub struct BackupOutcome {
-    pub path: PathBuf,
-    pub prune: BackupPruneOutcome,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Default)]
-pub struct BackupPruneOutcome {
-    pub kept_files: Vec<PathBuf>,
-    pub removed_files: Vec<PathBuf>,
-    pub error: Option<BackupError>,
-}
-
-#[allow(dead_code)]
-#[derive(Debug)]
-pub enum BackupError {
-    Io(io::Error),
-    Sqlx(sqlx::Error),
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum BackupRequestErrorKind {
-    ShutdownSkip,
-    Failure,
-}
-
-#[derive(Debug)]
-pub enum BackupRequestError {
-    ShutdownInProgress,
-    MissingHomeDirectory,
-    BackupFailed(BackupError),
-    ShutdownTimedOut,
-}
-
-impl fmt::Display for BackupError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Io(error) => write!(formatter, "{error}"),
-            Self::Sqlx(error) => write!(formatter, "{error}"),
-        }
-    }
-}
-
-impl std::error::Error for BackupError {}
-
-impl From<io::Error> for BackupError {
-    fn from(error: io::Error) -> Self {
-        Self::Io(error)
-    }
-}
-
-impl From<sqlx::Error> for BackupError {
-    fn from(error: sqlx::Error) -> Self {
-        Self::Sqlx(error)
-    }
-}
-
-impl BackupRequestError {
-    pub fn kind(&self) -> BackupRequestErrorKind {
-        match self {
-            Self::ShutdownInProgress => BackupRequestErrorKind::ShutdownSkip,
-            Self::MissingHomeDirectory
-            | Self::BackupFailed(_)
-            | Self::ShutdownTimedOut => BackupRequestErrorKind::Failure,
-        }
-    }
-}
-
-impl fmt::Display for BackupRequestError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::ShutdownInProgress => {
-                write!(formatter, "backup skipped because shutdown is in progress")
-            }
-            Self::MissingHomeDirectory => write!(formatter, "Cannot find home directory"),
-            Self::BackupFailed(error) => write!(formatter, "{error}"),
-            Self::ShutdownTimedOut => write!(formatter, "backup timed out during app shutdown"),
-        }
-    }
-}
-
-impl std::error::Error for BackupRequestError {}
-
-pub fn log_backup_request_error(context: &str, error: &BackupRequestError) {
-    match error.kind() {
-        BackupRequestErrorKind::ShutdownSkip => {
-            log::warn!("autobackup skipped after {context}: {error}");
-        }
-        BackupRequestErrorKind::Failure => {
-            log::error!("autobackup failed after {context}: {error}");
-        }
-    }
 }
 
 #[derive(Debug, Clone, Serialize)]
